deployment/Predict_Default_Payment_Credit_Card/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
pillow
scikit-learn